+++ src/main/java/mirsario/cameraoverhaul/CameraSystem.java
@@ -5,19 +5,19 @@ import mirsario.cameraoverhaul.utilities.*;
 import org.joml.*;
 
 public final class CameraSystem {
-	private float prevForwardVelocityPitchOffset;
-	private float prevVerticalVelocityPitchOffset;
-	private float prevStrafingRollOffset;
-	private float prevCameraYaw;
-	//private float prevYawDeltaRollOffset;
-	private float yawDeltaRollOffset;
-	private float yawDeltaRollTargetOffset;
+	private double prevForwardVelocityPitchOffset;
+	private double prevVerticalVelocityPitchOffset;
+	private double prevStrafingRollOffset;
+	private double prevCameraYaw;
+	//private double prevYawDeltaRollOffset;
+	private double yawDeltaRollOffset;
+	private double yawDeltaRollTargetOffset;
 	private final Transform offsetTransform = new Transform();
 
-	public void onCameraUpdate(CameraContext context, float deltaTime) {
+	public void onCameraUpdate(CameraContext context, double deltaTime) {
 		// Reset the offset transform
-		offsetTransform.position = new Vector3f(0, 0, 0);
-		offsetTransform.eulerRot = new Vector3f(0, 0, 0);
+		offsetTransform.position = new Vector3d(0, 0, 0);
+		offsetTransform.eulerRot = new Vector3d(0, 0, 0);
 
 		ConfigData config = CameraOverhaul.instance.config;
 
@@ -25,7 +25,7 @@ public final class CameraSystem {
 			return;
 		}
 
-		float strafingRollFactorToUse = config.strafingRollFactor;
+		double strafingRollFactorToUse = config.strafingRollFactor;
 		
 		if (context.isFlying) {
 			strafingRollFactorToUse = config.strafingRollFactorWhenFlying;
@@ -33,13 +33,13 @@ public final class CameraSystem {
 			strafingRollFactorToUse = config.strafingRollFactorWhenSwimming;
 		}
 
-		var relativeXZVelocity = VectorUtils.rotate(new Vector2f((float)context.velocity.x, (float)context.velocity.z), 360f - (float)context.transform.eulerRot.y);
+		var relativeXZVelocity = VectorUtils.rotate(new Vector2d((double)context.velocity.x, (double)context.velocity.z), 360d - (double)context.transform.eulerRot.y);
 
 		// X
 		verticalVelocityPitchOffset(context, offsetTransform, relativeXZVelocity, deltaTime, config.verticalVelocityPitchFactor, config.verticalVelocitySmoothingFactor);
 		forwardVelocityPitchOffset(context, offsetTransform, relativeXZVelocity, deltaTime, config.forwardVelocityPitchFactor, config.horizontalVelocitySmoothingFactor);
 		// Z
-		yawDeltaRollOffset(context, offsetTransform, relativeXZVelocity, deltaTime, config.yawDeltaRollFactor * 1.25f, config.yawDeltaSmoothingFactor, config.yawDeltaDecayFactor);
+		yawDeltaRollOffset(context, offsetTransform, relativeXZVelocity, deltaTime, config.yawDeltaRollFactor * 1.25d, config.yawDeltaSmoothingFactor, config.yawDeltaDecayFactor);
 		strafingRollOffset(context, offsetTransform, relativeXZVelocity, deltaTime, strafingRollFactorToUse, config.horizontalVelocitySmoothingFactor);
 
 		prevCameraYaw = context.transform.eulerRot.y;
@@ -50,29 +50,29 @@ public final class CameraSystem {
 		transform.eulerRot.add(offsetTransform.eulerRot);
 	}
 
-	private void verticalVelocityPitchOffset(CameraContext context, Transform outputTransform,  Vector2f relativeXZVelocity, float deltaTime, float intensity, float smoothing) {
-		float targetVerticalVelocityPitchOffset = context.velocity.y * 2.75f * (context.velocity.y < 0f ? 2.25f : 2.0f);
+	private void verticalVelocityPitchOffset(CameraContext context, Transform outputTransform,  Vector2d relativeXZVelocity, double deltaTime, double intensity, double smoothing) {
+		double targetVerticalVelocityPitchOffset = context.velocity.y * 2.75d * (context.velocity.y < 0d ? 2.25d : 2.0d);
 
-		if (context.velocity.y < 0f) {
-			targetVerticalVelocityPitchOffset *= 2.25f;
+		if (context.velocity.y < 0d) {
+			targetVerticalVelocityPitchOffset *= 2.25d;
 		}
 
-		float currentVerticalVelocityPitchOffset = (float) MathUtils.damp(prevVerticalVelocityPitchOffset, targetVerticalVelocityPitchOffset, smoothing, deltaTime);
+		double currentVerticalVelocityPitchOffset = (double) MathUtils.damp(prevVerticalVelocityPitchOffset, targetVerticalVelocityPitchOffset, smoothing, deltaTime);
 		
 		outputTransform.eulerRot.x += currentVerticalVelocityPitchOffset * intensity;
 		prevVerticalVelocityPitchOffset = currentVerticalVelocityPitchOffset;
 	}
 
-	private void forwardVelocityPitchOffset(CameraContext context, Transform outputTransform, Vector2f relativeXZVelocity, float deltaTime, float intensity, float smoothing) {
-		float targetForwardVelocityPitchOffset = relativeXZVelocity.y * 5f;
-		float currentForwardVelocityPitchOffset = (float)MathUtils.damp(prevForwardVelocityPitchOffset, targetForwardVelocityPitchOffset, smoothing, deltaTime);
+	private void forwardVelocityPitchOffset(CameraContext context, Transform outputTransform, Vector2d relativeXZVelocity, double deltaTime, double intensity, double smoothing) {
+		double targetForwardVelocityPitchOffset = relativeXZVelocity.y * 5d;
+		double currentForwardVelocityPitchOffset = (double)MathUtils.damp(prevForwardVelocityPitchOffset, targetForwardVelocityPitchOffset, smoothing, deltaTime);
 		
 		outputTransform.eulerRot.x += currentForwardVelocityPitchOffset * intensity;
 		prevForwardVelocityPitchOffset = currentForwardVelocityPitchOffset;
 	}
 
-	private void yawDeltaRollOffset(CameraContext context, Transform outputTransform, Vector2f relativeXZVelocity, float deltaTime, float intensity, float offsetSmoothing, float decaySmoothing) {
-		float yawDelta = prevCameraYaw - context.transform.eulerRot.y;
+	private void yawDeltaRollOffset(CameraContext context, Transform outputTransform, Vector2d relativeXZVelocity, double deltaTime, double intensity, double offsetSmoothing, double decaySmoothing) {
+		double yawDelta = prevCameraYaw - context.transform.eulerRot.y;
 
 		if (yawDelta > 180) {
 			yawDelta = 360 - yawDelta;
@@ -80,16 +80,16 @@ public final class CameraSystem {
 			yawDelta = -360 - yawDelta;
 		}
 
-		yawDeltaRollTargetOffset += yawDelta * 0.07f;
+		yawDeltaRollTargetOffset += yawDelta * 0.07d;
 		yawDeltaRollOffset = MathUtils.damp(yawDeltaRollOffset, yawDeltaRollTargetOffset, offsetSmoothing, deltaTime);
 
 		outputTransform.eulerRot.z += yawDeltaRollOffset * intensity;
 		
-		yawDeltaRollTargetOffset = MathUtils.damp(yawDeltaRollTargetOffset, 0f, decaySmoothing, deltaTime);
+		yawDeltaRollTargetOffset = MathUtils.damp(yawDeltaRollTargetOffset, 0d, decaySmoothing, deltaTime);
 	}
 
-	private void strafingRollOffset(CameraContext context, Transform outputTransform, Vector2f relativeXZVelocity, float deltaTime, float intensity, float smoothing) {
-		float strafingRollOffset = -relativeXZVelocity.x * 15f;
+	private void strafingRollOffset(CameraContext context, Transform outputTransform, Vector2d relativeXZVelocity, double deltaTime, double intensity, double smoothing) {
+		double strafingRollOffset = -relativeXZVelocity.x * 15d;
 		
 		prevStrafingRollOffset = strafingRollOffset = MathUtils.damp(prevStrafingRollOffset, strafingRollOffset, smoothing, deltaTime);
