+++ src/main/java/mirsario/cameraoverhaul/abstractions/MathAbstractions.java
@@ -13,11 +13,11 @@ import org.joml.*;
 
 public final class MathAbstractions {
 #if MC_RELEASE && MC_VERSION >= "11500"
-	public static void rotateMatrixByAxis(com.mojang.blaze3d.vertex.PoseStack matrix, float axisX, float axisY, float axisZ, float rotation) {
+	public static void rotateMatrixByAxis(com.mojang.blaze3d.vertex.PoseStack matrix, double axisX, double axisY, double axisZ, double rotation) {
 #if MC_VERSION >= "11903"
-		matrix.mulPose(Axis.of(new Vector3f(axisX, axisY, axisZ)).rotationDegrees(rotation));
+		matrix.mulPose(Axis.of(new Vector3f((float)axisX, (float)axisY, (float)axisZ)).rotationDegrees((float)rotation));
 #else
-		matrix.mulPose(new Vector3f(axisX, axisY, axisZ).rotationDegrees(rotation));
+		matrix.mulPose(new Vector3f((float)axisX, (float)axisY, (float)axisZ).rotationDegrees((float)rotation));
 #endif
 	}
 #endif
